+++ tests/benchmarks/computer_use/osworld/test_osworld_mediator.py
@@ -13,7 +13,7 @@ from agential.benchmarks.computer_use.osworld.osworld_mediator import OSWorldMed
 from agential.benchmarks.computer_use.osworld.osworld_data_loader import (
     OSWorldDataLoader,
 )
-from agential.benchmarks.computer_use.osworld.osworld_benchmark import OSWorldBenchmark
+from agential.benchmarks.computer_use.osworld.osworld import OSWorldBenchmark
 
 EXAMPLES_DIR: str = (
     "agential/benchmarks/computer_use/osworld/evaluation_examples/examples"
